Feature/add tests

+++ src/main/java/adelina/luxtravel/service/TravelingDataService.java
@@ -2,6 +2,7 @@ package adelina.luxtravel.service;
 
 import adelina.luxtravel.domain.*;
 import adelina.luxtravel.domain.transport.Transport;
+import adelina.luxtravel.domain.transport.TransportClass;
 import adelina.luxtravel.domain.wrapper.*;
 import adelina.luxtravel.exception.*;
 import adelina.luxtravel.repository.*;
@@ -9,6 +10,7 @@ import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -17,6 +19,7 @@ import java.util.Optional;
 import static adelina.luxtravel.utility.Constants.INVALID_ID;
 
 @Service
+@Transactional
 public class TravelingDataService {
     private TravelingDataRepository travelingDataRepository;
     private TransportRepository transportRepository;
@@ -31,9 +34,10 @@ public class TravelingDataService {
         this.travelingPointRepository = travelingPointRepository;
     }
 
+
     public TravelingData save(TravelingData travelingData) {
-        validateBookingData(travelingData);
-        validateFieldsExist(travelingData.getDepartureDestination(), travelingData.getTransport());
+        validateTravelingData(travelingData);
+
         return travelingDataRepository.save(travelingData);
     }
 
@@ -42,15 +46,18 @@ public class TravelingDataService {
             throw new InvalidArgumentException(INVALID_ID);
         }
 
-        Optional<TravelingData> bookingData = travelingDataRepository.findById(id);
+        Optional<TravelingData> travelingData = travelingDataRepository.findById(id);
 
-        if (!bookingData.isPresent()) {
-            throw new NonExistentItemException("This booking data does not exist");
+        if (!travelingData.isPresent()) {
+            throw new NonExistentItemException("This traveling data does not exist");
         }
-        return bookingData.get();
+        return travelingData.get();
     }
 
-    public List<TravelingData> findByDates(LocalDate from, LocalDate to) {
+    public List<TravelingData> findByDates(Date dates) {
+        LocalDate from = dates.getFromDate();
+        LocalDate to = dates.getToDate();
+
         validateDates(from, to);
 
         List<TravelingData> travelingData = travelingDataRepository.findByDates(from, to);
@@ -65,38 +72,57 @@ public class TravelingDataService {
         List<TravelingData> travelingData = travelingDataRepository.findAll();
 
         if (ObjectUtils.isEmpty(travelingData)) {
-            throw new NonExistentItemException("No booking data found");
+            throw new NonExistentItemException("No traveling data found");
         }
         return travelingData;
     }
 
-    public void updateTransport(long travelingDataId, Transport transport) {
-        if (travelingDataId <= NumberUtils.LONG_ZERO) {
-            throw new InvalidArgumentException("Invalid traveling data id");
-        }
+    public void updateDates(long travelingDataId, Date dates) {
+        validateUpdateDatesParameters(travelingDataId, dates);
 
-        findById(travelingDataId);
-        validateTransportData(transport);
+        LocalDate fromDate = dates.getFromDate();
+        LocalDate toDate = dates.getToDate();
 
-        long transportId = transport.getId();
-
-        travelingDataRepository.updateTransport(transportId, travelingDataId);
+        travelingDataRepository.updateFromDate(fromDate, travelingDataId);
+        travelingDataRepository.updateToDate(toDate, travelingDataId);
     }
 
-    public void deleteById(long id) throws InvalidArgumentException {
+    public void deleteById(long id) {
         if (id <= NumberUtils.LONG_ZERO) {
             throw new InvalidArgumentException(INVALID_ID);
+        }
+
+        if (!travelingDataExists(id)) {
+            throw new NonExistentItemException("Traveling data with this id does not exist");
         }
         travelingDataRepository.deleteById(id);
     }
 
-    private void validateBookingData(TravelingData travelingData) {
-        if (travelingData == null) {
-            throw new InvalidArgumentException("Invalid booking data");
+    private void validateFields(DepartureDestination departureDestination, Transport transport, Date date) {
+        if (departureDestination == null) {
+            throw new InvalidArgumentException("Invalid departure destination");
+        }
+
+        if (date == null) {
+            throw new InvalidArgumentException("Invalid dates");
+        }
+
+        validateTransport(transport);
+        validateTransportExists(transport);
+        validateDepartureDestinationExists(departureDestination);
+        validateDates(date.getFromDate(), date.getToDate());
+    }
+
+    private void validateTransportExists(Transport transport) {
+        long transportId = transport.getId();
+        Optional<Transport> transportOptional = transportRepository.findById(transportId);
+
+        if (!transportOptional.isPresent()) {
+            throw new NonExistentItemException("This transport does not exist");
         }
     }
 
-    private void validateFieldsExist(DepartureDestination departureDestination, Transport transport) {
+    private void validateDepartureDestinationExists(DepartureDestination departureDestination) {
         TravelingPoint departurePoint = departureDestination.getDeparturePoint();
         TravelingPoint destinationPoint = departureDestination.getDestinationPoint();
         long departurePointId = departurePoint.getId();
@@ -110,27 +136,65 @@ public class TravelingDataService {
         if (!endingPoint.isPresent()) {
             throw new NonExistentItemException("Destination traveling point does not exist");
         }
-
-        validateTransportData(transport);
     }
 
-    private void validateTransportData(Transport transport) {
+    private void validateUpdateDatesParameters(long travelingDataId, Date newDates) {
+        if (travelingDataId <= NumberUtils.INTEGER_ZERO) {
+            throw new InvalidArgumentException("Invalid traveling data id");
+        }
+
+        if (newDates == null) {
+            throw new InvalidArgumentException("Invalid from and to dates");
+        }
+
+        TravelingData travelingData = findById(travelingDataId);
+
+        validateDates(newDates.getFromDate(), newDates.getToDate());
+
+        if (newDates.equals(travelingData.getDate())) {
+            throw new AlreadyExistingItemException("New dates can not be the same as the current");
+        }
+    }
+
+    private boolean travelingDataExists(long travelingDataId) {
+        Optional<TravelingData> travelingData = travelingDataRepository.findById(travelingDataId);
+
+        return travelingData.isPresent();
+    }
+
+    private void validateDates(LocalDate from, LocalDate to) {
+        if (from == null || to == null ||
+                from.isEqual(to) || from.isAfter(to) || from.isBefore(LocalDate.now())) {
+            throw new InvalidArgumentException("Invalid from and to dates");
+        }
+    }
+
+    private void validateTransport(Transport transport) {
         if (transport == null) {
             throw new InvalidArgumentException("Invalid transport");
         }
 
-        long transportId = transport.getId();
-        Optional<Transport> foundTransport = transportRepository.findById(transportId);
+        TransportClass transportClass = transport.getTransportClass();
 
-        if (!foundTransport.isPresent()) {
-            throw new NonExistentItemException("This transport does not exist");
+        if (transportClass == null) {
+            throw new InvalidArgumentException("Invalid transport class");
         }
     }
 
-    private void validateDates(LocalDate from, LocalDate to) {
-        if (from == null || to == null || from.isEqual(to)
-                || from.isAfter(to) || from.isBefore(LocalDate.now())) {
-            throw new InvalidArgumentException("Invalid dates");
+    private void validateTravelingData(TravelingData travelingData) {
+        if (travelingData == null) {
+            throw new InvalidArgumentException("Invalid traveling data");
+        }
+
+        Transport transport = travelingData.getTransport();
+        DepartureDestination departureDestination = travelingData.getDepartureDestination();
+        Date date = travelingData.getDate();
+        int availableTicketsCount = travelingData.getAvailableTicketsCount();
+
+        validateFields(departureDestination, transport, date);
+
+        if (availableTicketsCount <= NumberUtils.INTEGER_ZERO) {
+            throw new InvalidArgumentException("Invalid tickets count");
         }
     }
 }
