+++ src/main/java/adelina/luxtravel/service/TravelingPointService.java
@@ -1,8 +1,7 @@
 package adelina.luxtravel.service;
 
 import adelina.luxtravel.domain.TravelingPoint;
-import adelina.luxtravel.exception.AlreadyExistingItemException;
-import adelina.luxtravel.exception.NonExistentItemException;
+import adelina.luxtravel.exception.*;
 import adelina.luxtravel.repository.TravelingPointRepository;
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -10,6 +9,7 @@ import org.apache.commons.lang3.math.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import adelina.luxtravel.exception.InvalidArgumentException;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,6 +17,7 @@ import java.util.Optional;
 import static adelina.luxtravel.utility.Constants.*;
 
 @Service
+@Transactional
 public class TravelingPointService {
     private TravelingPointRepository travelingPointRepository;
 
@@ -27,13 +28,14 @@ public class TravelingPointService {
 
     public TravelingPoint save(TravelingPoint travelingPoint) {
         validateTravelingPoint(travelingPoint);
-        validateTravelingPointDoesNotExist(travelingPoint);
+        validateTravelingPointDoesNotExist(travelingPoint.getName());
 
         return travelingPointRepository.save(travelingPoint);
     }
 
     public List<TravelingPoint> saveAll(List<TravelingPoint> travelingPoints) {
         validateTravelingPointsList(travelingPoints);
+
         return travelingPointRepository.saveAll(travelingPoints);
     }
 
@@ -47,6 +49,7 @@ public class TravelingPointService {
         if (!travelingPoint.isPresent()) {
             throw new NonExistentItemException("Traveling point with this id does not exist");
         }
+
         return travelingPoint.get();
     }
 
@@ -54,7 +57,14 @@ public class TravelingPointService {
         if (StringUtils.isEmpty(name)) {
             throw new InvalidArgumentException("Invalid name");
         }
-        return travelingPointRepository.findByName(name);
+
+        Optional<TravelingPoint> travelingPoint = travelingPointRepository.findByName(name);
+
+        if (!travelingPoint.isPresent()) {
+            throw new NonExistentItemException("Traveling point with this name does not exists");
+        }
+
+        return travelingPoint.get();
     }
 
     public List<TravelingPoint> findAll() {
@@ -63,21 +73,13 @@ public class TravelingPointService {
         if (ObjectUtils.isEmpty(travelingPoints)) {
             throw new NonExistentItemException("There are no traveling points found");
         }
+
         return travelingPoints;
     }
 
     public void updateName(String newName, String oldName) {
-        if (StringUtils.isEmpty(newName)) {
-            throw new InvalidArgumentException("Invalid new name");
-        }
-        if (StringUtils.isEmpty(oldName)) {
-            throw new InvalidArgumentException("Invalid old name");
-        }
-        if (newName.equals(oldName)) {
-            throw new InvalidArgumentException("New name can not be the same as the old one");
-        }
+        validateUpdateNameParameters(newName, oldName);
 
-        validateTravelingPointDoesNotExist(findByName(newName));
         travelingPointRepository.updateName(newName, oldName);
     }
 
@@ -85,6 +87,13 @@ public class TravelingPointService {
         if (id <= NumberUtils.LONG_ZERO) {
             throw new InvalidArgumentException(INVALID_ID);
         }
+
+        Optional<TravelingPoint> travelingPoint = travelingPointRepository.findById(id);
+
+        if (!travelingPoint.isPresent()) {
+            throw new NonExistentItemException("Traveling point with this id does not exist");
+        }
+
         travelingPointRepository.deleteById(id);
     }
 
@@ -92,8 +101,10 @@ public class TravelingPointService {
         if (ObjectUtils.isEmpty(travelingPoints)) {
             throw new InvalidArgumentException("Invalid list of traveling points");
         }
+
         for (TravelingPoint travelingPoint : travelingPoints) {
             validateTravelingPoint(travelingPoint);
+            validateTravelingPointDoesNotExist(travelingPoint.getName());
         }
     }
 
@@ -103,17 +114,25 @@ public class TravelingPointService {
         }
     }
 
-    private void validateTravelingPointDoesNotExist(TravelingPoint travelingPoint) {
-        String name = travelingPoint.getName();
-        double latitude = travelingPoint.getLatitude();
-        double longitude = travelingPoint.getLongitude();
+    private void validateTravelingPointDoesNotExist(String travelingPointName) {
+        Optional<TravelingPoint> travelingPoint = travelingPointRepository.findByName(travelingPointName);
 
-        TravelingPoint existingTravelingPoint = findByName(name);
-
-        if (name.equals(existingTravelingPoint.getName())
-                && latitude == existingTravelingPoint.getLatitude()
-                && longitude == existingTravelingPoint.getLongitude()) {
-            throw new AlreadyExistingItemException("This traveling point already exists");
+        if (travelingPoint.isPresent()) {
+            throw new AlreadyExistingItemException("Traveling point already exists");
         }
+    }
+
+    private void validateUpdateNameParameters(String newName, String oldName) {
+        if (StringUtils.isEmpty(newName)) {
+            throw new InvalidArgumentException("Invalid new name");
+        }
+        if (StringUtils.isEmpty(oldName)) {
+            throw new InvalidArgumentException("Invalid old name");
+        }
+        if (newName.equals(oldName)) {
+            throw new AlreadyExistingItemException("Traveling point with given new name already exists");
+        }
+
+        validateTravelingPointDoesNotExist(newName);
     }
 }
